# PERF: Use fastpath for accessing option value in internals