@jbrockmendel 2022-11-18
hmm at one point we had a code check that checked for private imports.  guess that got lost somewhere along the way