@WillAyd 2022-11-21
Curious if you've looked at making `get_option` a class and using `__slots__` instead for this? Seems like it would be the idiomatic approach to speeding up attribute access