@jorisvandenbossche 2022-11-24
@WillAyd you mean the `DictWrapper` class that `get_option`/`options` uses? That is currently using a dynamic `__getattr__` to look up attributes in a global dictionary that actually stores the option values. 
What you suggest might be an option (but not directly an idea, didn't look into that), but that seems a bigger change in the internals of the options machinery. While what I am doing here is just using the fact that the current options internals is just a plain dictionary.