@jbrockmendel 2022-11-28
Best Practices-wise, the best-cast would be to try to optimize get_option.  Next-best would be to make _global_config public-for-inernal-usage and canonize this as how we do option checks within pandas