@jorisvandenbossche 2022-11-28
> best-cast would be to try to optimize get_option

That would require a change of (or addition to) how `get_option` works, I think. Currently that takes a string (eg `get_option("mode.copy_on_write")`), and thus that requires parsing the string, and also has some overhead. 

It's easy to rename `_global_config` to `global_config` (its module is already private) if that is preferred.