@jbrockmendel 2022-11-28
yah even if we did optimize get_option a dict lookup will always be way faster